Refactor: Make `StoreCreationCountry` its own class for re-usability.
